Spell.Target is always lowercase, fix a few issues with animist's turrets
# Conflicts:
#	GameServer/ai/brain/StandardMobBrain.cs

+++ GameServer/gameobjects/GameNPC.cs
@@ -1358,7 +1358,7 @@ namespace DOL.GS
 		/// </summary>
 		public virtual void WalkToSpawn()
 		{
-			WalkToSpawn((short)(50));
+			WalkToSpawn(50);
 		}
 
 		/// <summary>
